[LibOS] Replace DENTRY_SYNTHETIC with a separate filesystem
Previously, synthetic files were created as special dentries with
DENTRY_SYNTHETIC flag set. These dentries were marked as belonging to a
filesystem (typically "chroot") but weren't really initialized by it. As
a result, they could break assumptions made by that filesystem (e.g. about
`dent->inode` existing).

This change treats synthetic files same as other special files (named
pipes and sockets). We set `dent->fs` to mark them as belonging to a
separate filesystem, and all operations for the file are forwarded to
that filesystem.

In addition, for the only type of synthetic files we create are
directories. If the user mounts something else than a directory, the
synthetic mountpoint can still be a directory: it will be shadowed by
the mounted file anyway.

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>